Code Refactor
This is a significant refactor. Cleaned up and simplified many pieces of
the code and as a result, near 100% code coverage.

Removed `ami` sub-package. It makes a lot more sense for the Image type
to be in the amicache package directly. Also renamed the type from AMI
to Image.

Overhauled `amicache`. There is now just an internal memory cache. The
cache itself is its own manager now.

The `api` package has also changed. There's no longer a custom http
Handler type. The `v1` package has been renamed to `query` and it
implements the http.Handler interface itself. As such, the cache is
provided directly instead of through the context values. Makes it
type safe and easier to test.

The main program now intercepts sigint and sigterm and terminates the
services gracefully.

In addition to caching ther region and image details for an AMI, it now
also stores the launch permissions assoicated with it. This does add a
significant load time when caching a large number of images in a single
account (> ~150 images). As such, there are two new caching tunables.
The README has more information.